pyproject.toml
[project]
name = "zm00"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kenji Sato", email = "mail@kenjisato.jp" }
]
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "lxml>=6.0.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
    "typer>=0.16.0",
]

[project.scripts]
zm00 = "zm00.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
